removed mutex lock to avoid race condition. Doubt the lock serves a purpose.
I don't think the mutex needs to be locked in the telemetry loop, since it is a read-only thread for that data. A race condition can occur since the sleep(1) is within the lock scope and immediately followed by continue and relock.
Since the other threads in writing can only write one value, a lock is likely not needed there either.