Ensure that no callbacks happen after endpoint is destroyed.
A number of different async callbacks (timers and socket
read/write/accept notfications) were occurring after the client
and server were destroyed. This is problematic because connections
have references to loggers owned by the endpoint and were referencing
them after they were destroyed.

The server was also having its handle_accept method called with
a "cancelled" status after it was destroyed.

This change ensures that all of the endpoints callbacks are cancelled
before members they reference are destroyed, and it also makes the
connections hang onto a reference to the endpoint, ensuring that it
does not get destroyed before they do.