add a workaround to mingw / libstdc++ issue 68307
This is a workaround to some standard non-conformance defect in
libstdc++ headers for mingw. Although the C++11 standard requires
that `std::errc::operation_canceled` is an enumerator value of
`std::errc` in `<system_error>`, and asio library uses std::errc
when compiled as standalone, mingw apparently doesn't use all the
std::errc signals, and doesn't define any of the enumerators that
it doesn't use.

Presumably mingw instead packs these signals into some different
error codes or something, it's not entirely clear. Regardless,
even if there are some low-level issues in mingw implementations
of sockets, websocketpp is a bit higher level library and one
would think that we should be able to try to run websocketpp over
any given sockets implementation and at least get it to compile.

This patch is a polyfill so that if for whatever reason the host
system fails to define this error code symbol, websocketpp will
just return false whenever testing if an error code matches it,
as a form of special support for mingw. (Or any other toolchain
that happens to have a similar bug -- there is no ifdef going on
here.)

In preliminary tests, websocketpp examples seem to work fine with
this patch when compiled with the mingw-w64 toolchain from ubuntu
trusty repo, with Asio standalone and C++11 standard, when
running in wine on my machine.