examples: make debug_client handle both secure and non-secure connections
This demonstrates how to flexibly support both secure and insecure connections
at runtime. It makes use of templates to differentiate and wrap the
incompatible classes and members. The result is as much common code as
possible.

Furthermore, the actually used URI and method are printed.

This change does not cover proper configuration of the TLS connection. A
proper implementation requires more code for choice of protocols,
verification of certificates, and so on.

Most notably, the given default URI wss://echo.websocket.org currently does
*not* even work with this example, because it makes use of a wildcard
certificate. This requires more asio boilerplate.