Update to make WebSocket++ work with modern Asio/Boost
Back in 2018, a bunch of Asio stuff was deprecated in Asio 1.12.0/Boost 1.66.0 to reflect the Networking TS.  Now that it’s been a while, many of these deprecated functions and interfaces have been removed.  As of Asio 1.34.0/Boost 1.87.0, WebSocket++ will not function.  This commit makes the following changes and allows for functionality modern versions of Asio/Boost:
 * Use `io_context` instead of `io_service`
 * Use `basic_waitable_timer::expiry` instead of `basic_waitable_timer::expires_from_now`
 * Call `asio::post` instead of `io_context::post`
 * Use `executor_work_guard` instead of `io_context::work`
 * Use `resolver::resolve` directly instead of creating `resolver::query`
 * Use `resolver::results_type` instead of `resolver::iterator directly`
 * Use `max_listen_connextions` instead of `max_connections`
 * Use `io_context::restart` instead of `io_context::reset`
 * Use `make_address` instead of `from_string`